Add a script to generate the wiki's spell modules
This generates Module:Table of spells and Module:Table of spellbooks.
The output is in two files, wiki_spell_table.txt and
wiki_spellbook_table.txt, which can be copied over to the wiki.

The script runs on game load - to run the script, compile Crawl
on this branch, run the game, and start a new game of crawl or load
up an old save file (either works).